// src/app/models/chars.model.ts
export class Chars {
  update(id: any, data: { name: string | undefined; status: string | undefined; species: string | undefined; gender: string | undefined; origin: string | undefined; location: string | undefined; imagen: any; }) {
    throw new Error('Method not implemented.');
  }
  id?: any;
  name?: string;
  status?: string;
  species?: string;
  gender?: string;
  origin?: string;
  location?: string;
  imagen?: any;
}